Make items a sub-module to keep the large list of things relatively separate
