+++ eldritch/items/spells.py
@@ -1,281 +1,8 @@
-import operator
-
-from eldritch.assets import Card, Deputy
 from eldritch import events
+from eldritch import monsters
 from eldritch import places
 from eldritch import values
-from eldritch import monsters
-
-
-class Item(Card):
-
-  ITEM_TYPES = {"weapon", "tome", None}
-
-  def __init__(
-      self, name, idx, deck, active_bonuses, passive_bonuses, hands, price, item_type=None,
-  ):
-    assert item_type in self.ITEM_TYPES
-    super().__init__(name, idx, deck, active_bonuses, passive_bonuses)
-    self.hands = hands
-    self.price = price
-    self.item_type = item_type
-
-  def hands_used(self):
-    return self.hands if self.active else 0
-
-
-class Weapon(Item):
-
-  def __init__(self, name, idx, deck, active_bonuses, passive_bonuses, hands, price):
-    super().__init__(name, idx, deck, active_bonuses, passive_bonuses, hands, price, "weapon")
-
-
-class OneshotWeapon(Weapon):
-
-  def get_trigger(self, event, owner, state):
-    if not isinstance(event, events.Check) or event.check_type != "combat":
-      return None
-    if event.character != owner or not self.active:
-      return None
-    return events.DiscardSpecific(event.character, [self])
-
-
-class Food(Item):
-
-  def __init__(self, idx):
-    super().__init__("Food", idx, "common", {}, {}, None, 1)
-
-  def get_usable_interrupt(self, event, owner, state):
-    if not isinstance(event, events.GainOrLoss) or owner != event.character:
-      return None
-    if isinstance(event.losses.get("stamina"), values.Value):
-      if event.losses["stamina"].value(state) < 1:
-        return None
-    elif event.losses.get("stamina", 0) < 1:
-      return None
-
-    discard = events.DiscardSpecific(event.character, [self])
-    prevent = events.LossPrevention(self, event, "stamina", 1)
-    return events.Sequence([discard, prevent], owner)
-
-
-class Whiskey(Item):
-
-  def __init__(self, idx):
-    super().__init__("Whiskey", idx, "common", {}, {}, None, 1)
-
-  def get_usable_interrupt(self, event, owner, state):
-    if not isinstance(event, events.GainOrLoss) or owner != event.character:
-      return None
-    if isinstance(event.losses.get("sanity"), values.Value):
-      if event.losses["sanity"].value(state) < 1:
-        return None
-    elif event.losses.get("sanity", 0) < 1:
-      return None
-
-    discard = events.DiscardSpecific(event.character, [self])
-    prevent = events.LossPrevention(self, event, "sanity", 1)
-    return events.Sequence([discard, prevent], owner)
-
-
-class ResearchMaterials(Item):
-
-  def __init__(self, idx):
-    super().__init__("Research Materials", idx, "common", {}, {}, None, 1)
-
-  def get_spend_amount(self, event, owner, state):
-    if not isinstance(event, events.SpendMixin) or event.is_done():
-      return None
-    if event.character != owner or "clues" not in event.spendable:
-      return None
-    if self.handle in event.spent_handles():
-      return False
-    return {"clues": 1}
-
-  def get_spend_event(self, owner):
-    return events.DiscardSpecific(owner, [self])
-
-
-class Axe(Weapon):
-
-  def __init__(self, idx):
-    super().__init__("Axe", idx, "common", {"physical": 2}, {}, 1, 3)
-    self._two_handed = False
-
-  def get_bonus(self, check_type, attributes):
-    bonus = super().get_bonus(check_type, attributes)
-    if check_type == "physical" and self._two_handed:
-      bonus += 1
-    return bonus
-
-  def deactivate(self):
-    self._two_handed = False
-
-  def hands_used(self):
-    return sum([self._active, self._two_handed])
-
-  def json_repr(self):
-    data = super().json_repr()
-    if self.active:
-      data["hands"] = self.hands_used()
-    return data
-
-
-class Bullwhip(Weapon):
-
-  def __init__(self, idx):
-    super().__init__("Bullwhip", idx, "common", {"physical": 1}, {}, 1, 2)
-
-  def get_usable_trigger(self, event, owner, state):
-    if not isinstance(event, events.Check) or owner != event.character:
-      return None
-    if event.check_type != "combat":
-      return None
-    return None  # TODO: create an event here
-
-
-class Cross(Weapon):
-
-  def __init__(self, idx):
-    super().__init__("Cross", idx, "common", {}, {"horror": 1}, 1, 3)
-
-  def get_bonus(self, check_type, attributes):
-    bonus = super().get_bonus(check_type, attributes)
-    if self.active and check_type == "magical" and attributes and "undead" in attributes:
-      bonus += 3
-    return bonus
-
-
-class Derringer18(Weapon):
-
-  def __init__(self, idx):
-    super().__init__(".18 Derringer", idx, "common", {"physical": 2}, {}, 1, 3)
-    self.losable = False
-
-
-class Tome(Item):
-
-  def __init__(self, name, idx, deck, price, movement_cost):
-    super().__init__(name, idx, deck, {}, {}, None, price, "tome")
-    self.movement_cost = movement_cost
-
-  def get_usable_interrupt(self, event, owner, state):
-    if not isinstance(event, events.CityMovement) or event.character != owner or event.is_done():
-      return None
-    if self.exhausted:
-      return None
-    if event.character.movement_points < self.movement_cost:
-      return None
-    return events.ReadTome([
-        events.ExhaustAsset(owner, self),
-        events.ChangeMovementPoints(owner, -self.movement_cost),
-        self.read_event(owner),
-    ], owner)
-
-  def get_usable_trigger(self, event, owner, state):
-    if not isinstance(event, events.WagonMove) or event.character != owner:
-      return None
-    if self.exhausted:
-      return None
-    if event.character.movement_points < self.movement_cost:
-      return None
-    return events.ReadTome([
-        events.ExhaustAsset(owner, self),
-        events.ChangeMovementPoints(owner, -self.movement_cost),
-        self.read_event(owner),
-    ], owner)
-
-  def read_event(self, owner):  # pylint: disable=unused-argument
-    return events.Nothing()
-
-
-class AncientTome(Tome):
-
-  def __init__(self, idx):
-    super().__init__("Ancient Tome", idx, "common", 4, 2)
-
-  def read_event(self, owner):
-    check = events.Check(owner, "lore", -1)
-    success = events.Sequence(
-        [events.Draw(owner, "spells", 1), events.DiscardSpecific(owner, [self])], owner,
-    )
-    return events.PassFail(owner, check, success, events.Nothing())
-
-
-def DarkCloak(idx):
-  return Item("Dark Cloak", idx, "common", {}, {"evade": 1}, None, 2)
-
-
-def Lantern(idx):
-  return Item("Lantern", idx, "common", {}, {"luck": 1}, None, 3)
-
-
-def Knife(idx):
-  return Weapon("Knife", idx, "common", {"physical": 1}, {}, 1, 2)
-
-
-def CavalrySaber(idx):
-  return Weapon("Cavalry Saber", idx, "common", {"physical": 2}, {}, 1, 3)
-
-
-def Revolver38(idx):
-  return Weapon(".38 Revolver", idx, "common", {"physical": 3}, {}, 1, 4)
-
-
-def Automatic45(idx):
-  return Weapon(".45 Automatic", idx, "common", {"physical": 4}, {}, 1, 5)
-
-
-def Shotgun(idx):  # NOTE: shotgun's special ability is hard-coded into character's count_successes
-  return Weapon("Shotgun", idx, "common", {"physical": 4}, {}, 2, 6)
-
-
-def Rifle(idx):
-  return Weapon("Rifle", idx, "common", {"physical": 5}, {}, 2, 6)
-
-
-def Dynamite(idx):
-  return OneshotWeapon("Dynamite", idx, "common", {"physical": 8}, {}, 2, 4)
-
-
-def HolyWater(idx):
-  return OneshotWeapon("Holy Water", idx, "unique", {"magical": 6}, {}, 2, 4)
-
-
-def EnchantedKnife(idx):
-  return Weapon("Enchanted Knife", idx, "unique", {"magical": 3}, {}, 1, 5)
-
-
-def EnchantedBlade(idx):
-  return Weapon("Enchanted Blade", idx, "unique", {"magical": 4}, {}, 1, 6)
-
-
-def MagicLamp(idx):
-  return Weapon("Magic Lamp", idx, "unique", {"magical": 5}, {}, 2, 7)
-
-
-class MagicPowder(Weapon):
-  def __init__(self, idx):
-    super().__init__("Magic Powder", idx, "unique", {"magical": 9}, {}, 2, 6)
-
-  def get_trigger(self, event, owner, state):
-    if not isinstance(event, events.Check) or event.check_type != "combat":
-      return None
-    if event.character != owner or not self.active:
-      return None
-    return events.Sequence([
-        events.DiscardSpecific(event.character, [self]),
-        events.Loss(event.character, {"sanity": 1}),
-    ], event.character
-    )
-
-
-def SwordOfGlory(idx):
-  return Weapon("Sword of Glory", idx, "unique", {"magical": 6}, {}, 2, 8)
-
-
-def TommyGun(idx):
-  return Weapon("Tommy Gun", idx, "common", {"physical": 6}, {}, 2, 7)
+from .base import Item
 
 
 class Spell(Item):
@@ -607,92 +334,3 @@ class FindGate(Spell):
           owner,
       )
     return events.Return(owner, owner.place.info.name)
-
-
-class DeputysRevolver(Weapon):
-
-  def __init__(self):
-    super().__init__("Deputy's Revolver", None, "tradables", {"physical": 3}, {}, 1, 0)
-    self.losable = False
-
-
-class PatrolWagon(Card):
-
-  def __init__(self):
-    super().__init__("Patrol Wagon", None, "tradables", {}, {})
-
-  def get_usable_interrupt(self, event, owner, state):
-    if not isinstance(event, events.CityMovement) or event.character != owner:
-      return None
-    if event.is_done() or event.moved:
-      return None
-    choice = events.PlaceChoice(owner, "Move using Patrol Wagon?", none_choice="Cancel")
-    # TODO: add an annotation to the PlaceChoice (e.g. "Move").
-    was_cancelled = values.Calculation(choice, None, operator.methodcaller("is_cancelled"))
-    patrol = events.ForceMovement(owner, choice)
-    cancel = events.CancelEvent(event)
-    move = events.WagonMove([patrol, cancel], owner)
-    cond = events.Conditional(owner, was_cancelled, None, {0: move, 1: events.Nothing()})
-    return events.Sequence([choice, cond], owner)
-
-  def get_trigger(self, event, owner, state):
-    if not isinstance(event, (events.Combat, events.Return)) or event.character != owner:
-      return None
-    die = events.DiceRoll(owner, 1)
-    discard = events.DiscardSpecific(owner, [self], to_box=True)
-    cond = events.Conditional(owner, values.Die(die), None, {0: discard, 2: events.Nothing()})
-    return events.Sequence([die, cond], owner)
-
-
-def CreateCommon():
-  common = []
-  for item in [
-      AncientTome, Automatic45, DarkCloak, Derringer18, Revolver38, Dynamite, Rifle, Shotgun,
-      TommyGun, Food, ResearchMaterials, Bullwhip, Cross, CavalrySaber, Knife, Whiskey, Axe,
-      Lantern,
-  ]:
-    common.extend([item(0), item(1)])
-  return common
-
-
-def CreateUnique():
-  counts = {
-      HolyWater: 4,
-      EnchantedBlade: 2,
-      EnchantedKnife: 2,
-      MagicLamp: 1,
-      MagicPowder: 2,
-      SwordOfGlory: 1,
-  }
-  uniques = []
-  for item, count in counts.items():
-    uniques.extend([item(idx) for idx in range(count)])
-  return uniques
-
-
-def CreateSpells():
-  counts = {
-      BindMonster: 2,
-      DreadCurse: 4,
-      EnchantWeapon: 3,
-      FindGate: 4,
-      FleshWard: 4,
-      # Heal: 3,  TODO: add heal to the spell pool when it does not stall upkeep in tests
-      Mists: 4,
-      RedSign: 2,
-      Shrivelling: 5,
-      Voice: 3,
-      Wither: 6,
-  }
-  spells = []
-  for item, count in counts.items():
-    spells.extend([item(idx) for idx in range(count)])
-  return spells
-
-
-def CreateTradables():
-  return [DeputysRevolver(), PatrolWagon()]
-
-
-def CreateSpecials():
-  return [Deputy()]
